+++ macaroons/index.js
@@ -1,0 +1,3 @@
+const grantAccess = require('./grant_access');
+
+module.exports = {grantAccess};